exe//pc8001/exprom1: Take over `MON`; alternate exit to original monitor
We have now configured the BASIC `MON` command to start tmon, instead
of the original ROM monitor. The original is still available by
specifying the `t1` parameter q`uit; ^B in the original monitor will
come back to tmon because that's generally less invasive and more
convenient.

We leave in place `CMD TMON` (because there are also other commands
under `CMD`) and the setting of all other Disk BASIC statement hooks
to enter tmon, which is still currently being used for debugging which
statements call which hooks.