src//tmon.i80: Checksum param area before quit
We were doing a checksum on the saved parameters area before executing
code (`k`all, `j`ump, etc.) so that they don't get re-initialised if
that code returns to the monitor, but were not doing this when we
quit. That meant that if we e.g. quit back to BASIC or a system
monitor and re-entered tmon, any parameter changes we made in the
previous session would be wiped out.

Now we do the checkum on quit as well, so that we can come back in
again to the same parameters we had when we left (assuming nobody
stomps on them of course).

Co-authored-by: Greg Bartholomaus <greg.bartholomaus@gmail.com>