WIP src/mc68/checksum/crc_16_ccitt: XXX
Added test_cksum_crc_16_ccitt()  (0xFFFF, []) test for both i8080 and
mc68. i8080 is broken because the loop is structured always to do one
update. Maybe we just ignore that for now?

Copied i8080 code, changing:
- HL (start of area to checksum) → X
- DE → `cksum` word in memory
- C (count) → B
- BC (count) → cklen word in memory

Feels like we need a `ckstart` as well, because of the awkwardness of
passing in X. But maybe not.

TODO:
- get cksum_crc_16_ccitt working for B bytes