i8080//crc_16_ccitt: Move to single API that takes long lengths
XXX For some reason testing up to length $DE00 works, but length $DF00
and above give a bad checksum error. We would expect this for a length
≥ $FF80, which would put test data on top of the routine, or maybe
even ≥ $FF00, which wraps around the beginning of memory, but not from
something lower than that. It's not clear if this is an error in the
test framework or if there really is something wrong with the code.

The `cksum_crc_16_ccitt` I stole from the Internet took the length in
C, using B for other purposes, limiting the maximum length of data to
sum to 256 bytes (0=256). To work around this, I added a second
routine, `lcksum_crc_16_ccitt`, that took a 16-bit length in BC and
called `ckupd_crc_16_ccitt` multiple times.

There's actually no need for this since we can pass a full 16-bit
length to `ckupd_crc_16_ccitt` and, with minor changes: a single push
and pull around the code that updates the checksum for a single
character, which needs B as a counter; and a change to a 16-bit rather
than 8-bit test at the end of the loop.

The comment about this being an "almost generic interface" etc. is
removed; it's kind of obvious, and the details it suggested about how
to do it were not necessarily correct.