src/crc_16_ccitt.py: Extract common test for crc_16_ccitt functions
This introduces a new technique to use a setup(m, …) function (not a
pytest fixture) from the importing module to set the parameters for
the routine and a result(m) function to return the results. This
extends our genericism to, e.g., depositing parameters to memory as
well as setting up registers.

As well as the existing verbosity of how we import these, this
introduces additional verbosity in the setup/result functions because
we must pass in the machine instance and the functions need to
explicitly get the Registers and SymTab objects from that. Perhaps
we can work out a better way of handling this.