src/i85/prstr: Generic prstr and prnl routines for platforms without them
These are generated only if `prstr` and `prnl`, respectively, are not
already defined, so that we don't have extra code (and a duplicate
symbol!) if they're already defined via calls to the platfrom BIOS or
whatever.

This changes several exe/* programs now to include i8080/prstr.i80
rather than have their own custom definitions of prstr and/or prnl.