mc68: Rename prstr0 → prstr; remove duplicate in generic "hello" program
A zero byte is pretty much the assumed terminator unless another is
explicitly specified so we make the mc68 `prstr0` routine name match
the `prstr` we use with other CPUs.

Remove the duplicate `prstr`/`prstr0` in the generic mc68 "hello"
program and instead include src/mc68/pr/str.a68 where we need it.
(We don't need it for e.g. the JR-200, which has this in the BIOS.)