src/jr200/biosdef: prstr0b → prstrQ
This prints 0-terminated strings, so we no longer need the 0 in the
name, per the last commit.

Make it more clear with a trailing `Q` that this prints only if
"quiet" mode is not set. I don't recall why I thought that a trailing
`b` was a clear way to state this.