exe/sharp/tmon48: Add init code before monitor start point
The monitor start continues to be at $C800 (so that it's easy to
remember); the init code is placed before that (with a check for
overflow) because it can be overwritten by BASIC or whatever else
happens to want to use that memory after tmon has been started once.

We also now use the `intentry` entry point so that we can see if any
loads of this happen via a `call`, but we've confirmed that the MZ-80K
and MZ-700 IPL ROMs do a `jp`, so this may want to be switched to the
`entry` entrypoint.