88 → 8: Rename and fix broken path that kept this from working
I've not been using this for a while, but I should be because it's
quite convenient, and messes up the environment less than directly
sourcing t8dev/t8dev/t8setup.bash.

I leave it at the top level, rather than in scripts/, to save typing
and to have a more obvious reminder that it's there.