i8080/clic/qtok: Token parsing (sym1/2, intrinsic consts, numbers)
This adds src/clic/grammar.md which does not yet discuss in detail
what our grammar is, but gives some thoughts on where things are
heading, along with references for and notes on Scheme and Common
Lisp. (We're tending towards the CL way of doing things at the
moment.)

The smallint code probably has quite a lot of room for optimisation:
it was hacked to work rather than be particularly clean.