exe/cpm/tmonlo: Fix checks for stomping on other code/data
The check to see if we were stomping on tmon code was wrong; we needed
to check that the initialisation code doesn't cross into tmon_rom, not
tmon_ram. (We don't care if it overlaps tmon_ram, since that will be
re-initialised after the init code has run anyway.)

Add another check to ensure that prvattr_* is not overlapping tmon_ram,
becuase if the tmon_ram init/usage stomps on prvattr_* then our terminal
codes will be destroyed.