src/i8080/tmon: Replace tmon_stksz with tmon_stack location
tmon previously always put it's stack at tmon_ram + tmon_stksz, with the
RAM work area just above that. But sometimes we need to put the stack
separate from the work area, and anyway it was confusing to try and
calculate where the work area is based on the stack size.

Now we put the stack at tmon_stack (growing down from the address
just below this), and the RAM work area starts at tmon_ram.