src/{tmon/test.py,i8080/tmon.pt}: Genericise calc (`/` command) test
We make this particularly simple by avoiding testing of any state,
just giving a command as typed in and comparing the output.

We obviously need to test values not specified (i.e., saved
parameters), but that can be done in a generic test of the parameter
system. We also for future tests will have to set up and/or check
memory state, which will provoke some refactoring, but we still intend
to move towards giving commands to parse, rather than lower-level
tests because that makes them more easily generic across different
CPUs.