WIP i8080/tmon intelhex: `:` command to read Intel hex records
XXX We have a naming issue here; some q* functions read input, others
just parse already-read input.

This turned out to be rather more difficult than originally thought
due to UI issues: since it's really for copy/paste or upload, rather
than interactive use, a few things needed to be reworked to handle
that, especially if you're pasting a large set of records and want to
know at the end how successful you've been. (The documentation
explains this.) The main refactoring here was to change `qhexdigitDE`
just to return carry set on error, leaving the caller to handle how it
wants to display the error.

There are reasonable checks in place to abort things if there's an
error (typically a non-hex digit) in the input. We do not currently
confirm the checksum, though.