exe/kc85/m100/tmon: Fix test for workspace overflow into stack
When we changed from having the stack always below the work area to
having it independent, and moved the stack to the top of meory (above
the work area), we didn't change the tests that checked for overflow
(over-allocation) of those areas. Fix this so we now define a minimum
stacksize and check that tmon_ram, below it, doesn't extend into that
minimum stack size. Also fix some broken comments related to this.