exe/sharp/tmon48: Make more room for code; improve overflow checking
Like the TRS-80 Model 100 verison, this now has a MINSTACKSZ
definition and check to ensure that tmon_ram doesn't overflow
into the reserved stack area above it.

We also move the code start down by 256 bytes because we had
less than 100 bytes of code space left when it was at $C800.