i8080/tmon qhexdigitDE: Set carry to indicate error status
Though not needed by `qhex`, future callers will need to be able to
check to see if a valid digit was read or not since they will want to
read fixed number of valid digits, rather than as many as the user
cares to type.

This adds only one byte to the code size.