i8080/tmon qhexdigitDE: Set carry to indicate error status, preserve B
Though not needed by `qhex`, future callers will need to be able to
check to see if a valid digit was read or not since they will want to
read fixed number of valid digits, rather than as many as the user
cares to type. Adds one byte to code size.

We now preserve B, as well as C and HL, which we also believe will
help future callers. Adds two bytes to code size.

Internally, the new rightmost digit is now saved in C instead of B,
and B is used (instead of the stack) for the a temp copy of the upper
nybble, because B,C ⇔ upper,lower is a bit more mnemonic.