i8080/tmon: Add ignored chars, and ^M → ^K for newline command
The common whitespace chars, space, CR and NL, are now in a new
category, "ignored" characters which do not generate an error as
they used to.

This helps when pasting in or uploading command sequences since you
can now have blank lines in them, and in the future it will avoid
errors or extra space being printed when a multi-line `:` Intel hex
paste/upload is done.