i8080/tmon read_params: Fix bug that broke good params after a bad one
During parameter entry for a command if one typed a bad parameter
(e.g., `t` when entering params for the examine command), one could
from that point on in entry of that command no longer enter a good
parameter (e.g., `s`).

It's not clear how this triggered the bug, but the bug itself was
clear: going through `.next` a second time HL might be set to
something other than the base of the params table, and the "ignore
spaces" branch was not jumping to `.next` but instead to to
`read_params`, which overwrote `t2` with the garbage HL, causing
`.parse_name` no longer to be able to load the params table.