exe/tmc/i8080/tmon: Move entry/exit code much further away from stack
When we changed from tmon_stksz to tmon_stack it re-ordered things and
we moved the stack at $200. Unfortunately, this gave us only a dozen
bytes of stack before starting to overwrite the tmc init/exit code.
This just barely worked with the current version of tmon, but caused
the exit code to be overwritten in development versions, so that
entering data with `,` caused the `q` command to fail. (Unless you set
the `t` parameter while using it; we don't know why that made it work.)

Move the entry/exit code down a lot, though still significantly above
the RAM workspace, so that we now have more than 128 bytes of stack.