tmon/test test_quit(): Moved from i8080/tmon.pt
This also tests certain characteristics of the param parser. This is
somewhat awkward; it might be worth looking at if there's a better way
to do this. But possibly the param parser is only independently
testable with machine-specific code.

The "call to `exit`" API has been changed to have the parameter value
in A, and this is tested in the machine-independent test_quit().

This leaves the machine-specific test_quit_default() and
test_quit_param() in place because the former also checks that the CRC
of the saved parameters is updated; this should be moved to the
machine-independent tests and then we can delete those
machine-dependent ones.

The 8080 code is updated to remove the `prompt.end` label (which is no
longer needed), allowing us to do a tail call optimization at the end
of `prompt.cmderr`. This saves a byte, but we added two due to the API
change above.