clic/objref: Rename; remove class; allow `str` for symbols
If a `str` is passed to sym1() or sym2(), it's converted to a `bytes`
using encoding 'ASCII'.

Putting these static functions into a class isn't necessary; the
module system itself provides all the namespace control we need.
(If a client wants a prefix, instead of `from … import *` they can
do an `import … as …`.