i8080/clic/rform: Reading tokens and forms
This reads a form, terminating on whitespace. Therefore `1 (+ 2 3) 3`
followed by a newline is three forms that require three calls to read.
Perhaps we want to read all three before executing them, as in an
implicit `progn` or `begin`? But that's for the evaluator to decide
once one or more of these are read into the heap.