exe//sd-sbc200/: Build "hello" program
The BIOS had to be split into two separate files: bios.inc with the
definitions and bioscode.inc with the generated code. This is because
the reset macro must be defined before it's used, but the serial I/O
code can't be at the start because you need to start with the reset
macro. (These files must not be named *.z80 because they cannot be
assembled standalone, and all files under exe/ ending in CPU
extensions are automatically assembled standalone.)

Note that while serial-echo uses (decimal) "13" for the timer/baud
rate, bioscode.inc uses (hex) "$0D". Having 13 as the only decimal
number in a program where everything else was hex led to a misreading
of that and incorrect setting of the baud rate in a previous version.

Build .bin files for both hello and serial-echo; the latter lets us
check our output against the original .bin file that came with the
source.