exe//sd-sbc200/: Build "hello" program
The BIOS had to be split into two separate files: bios.inc with the
definitions and bioscode.inc with the generated code. This is because
the reset macro must be defined before it's used, but the serial I/O
code can't be at the start because you need to start with the reset
macro.

Build .bin files for both hello and serial-echo; the latter lets us
check our output against the original .bin file that came with the
source.