Test: Copy only *.md from src/tmon/ to $reldir/tmon
We were previously copying all files, and this was causing an error
if there were any subdirectories there, such as the __pycache__/ that
appears if developers don't have $PYTHONPYCACHEPREFIX set.

But actually in the release ZIP we don't want even all the files in
there: just the documentation, not the Python code to support
testing. So copying *.md both fixes the error and gives us the
correct files for the release.

(Possibly we don't even want all the docs there. But that's a decision
for a different time.)