WIP exe/sbc/sd-sbc200/tmon.z80: tmon monitor and a small BIOS
ROM_BASE is $E000; RAM use is $FF00-$FFFF, which works with three
different ROM size configurations. See PLATFORM.md and exe//README.md
for details on this.

Test includes the exe//README.md in the release because we reference
it from PLATFORM.md. (The memory config is too complex to go into all
the details in PLATFORM.md,)

XXX `toml` instead of `tomllib` because latter doesn't support writing and
not in 3.9 or 3.10.