exe//sd-sbc200/hello: Set stack
This had been tested to work without setting the stack, but that was
probably just because we lucked out with the stack pointer being $0000
(or at least something not too unsuitable) after power-up. If the SP
happened to be a different value before a reset, this program would
crash when trying to return from bios_init.