t8dev,conf/t8t.conf,exe/: t8t config for a1 and sbc/aki80; sample hex upload
exe/sbc/aki80/chars.z80 is a small amount of data at $9000 that can be
uploaded using `t8t akimon send` to confirm that t8t and the config here
are working. See comments in that file for more details.

t8dev/ is updated to the latest `main`, f69450f.