exe//aki80/tmon-akimon: tmon for SuperAKI-80 SBC (RAM version)
This is designed to be uploaded in Intel HEX format using the
AKI-80モニター ROM `LH` command.

This also updates t8t.conf with a new 'aki' entry for Intel HEX
uploads to tmon. It needs a 5 ms character delay because it seems that
tmon is _really_ slow at processing this stuff, even on a 10 MHz Z80.
This appears as if it may be due to echoing all the input; turning off
echo for the hex bytes (but leaving it on for the prompt and `:`
command) fixes it for smaller files, but not larger ones. We'll have
to look at a fully non-echoed version (perhaps via the `y` command to
load data from the terminal to memory) to see if this fixes it.